Renamed gallery to AdventureGallery
